--- a/gaurun/global.go
+++ b/gaurun/global.go
@@ -4,16 +4,16 @@ import (
 	"net/http"
 	"sync"
 
-	"github.com/Sirupsen/logrus"
 	"github.com/mercari/gcm"
+	"github.com/uber-go/zap"
 )
 
 var (
 	ConfGaurun        ConfToml
 	QueueNotification chan RequestGaurunNotification
 	CertificatePemIos CertificatePem
-	LogAccess         *logrus.Logger
-	LogError          *logrus.Logger
+	LogAccess         zap.Logger
+	LogError          zap.Logger
 	StatGaurun        StatApp
 	// for numbering push
 	OnceNumbering sync.Once