--- a/glide.lock
+++ b/glide.lock
@@ -1,5 +1,5 @@
-hash: 8bf57a9cc5db0bc0cca2797499a6a3754733e3841d9ec2bcf72f74f6ce6364fa
-updated: 2016-06-13T11:53:25.375685652+09:00
+hash: a6443b44da0790750567bba0e3bba4887c8c4c7f1f90e7c014f970bffabbd2fb
+updated: 2016-08-28T19:44:15.038360984+09:00
 imports:
 - name: github.com/BurntSushi/toml
   version: bbd5bb678321a0d6e58f1099321dfa73391c1b6f
@@ -13,17 +13,27 @@ imports:
   - payload
   - payload/badge
   - push
-- name: github.com/Sirupsen/logrus
-  version: 4b6ea7319e214d98c938f12692336f7ca9348d6b
 - name: github.com/stretchr/testify
   version: f390dcf405f7b83c997eac1b06768bb9f44dec18
+  subpackages:
+  - assert
+  - suite
+  - require
+- name: github.com/uber-go/atomic
+  version: 0c9e689d64f004564b79d9a663634756df322902
+- name: github.com/uber-go/zap
+  version: d11d2851fcabcf03421c4dbdbc3146fc74eb1035
 - name: golang.org/x/net
   version: b797637b7aeeed133049c7281bfa31dcc9ca42d6
   subpackages:
   - http2
   - http2/hpack
-- name: golang.org/x/sys
-  version: f64b50fbea64174967a8882830d621a18ee1548e
+testImports:
+- name: github.com/davecgh/go-spew
+  version: 2df174808ee097f90d259e432cc04442cf60be21
   subpackages:
-  - unix
-devImports: []
+  - spew
+- name: github.com/pmezard/go-difflib
+  version: d8ed2627bdf02c080bf22230dbb337003b7aba2d
+  subpackages:
+  - difflib