--- a/glide.lock
+++ b/glide.lock
@@ -1,5 +1,5 @@
-hash: a6443b44da0790750567bba0e3bba4887c8c4c7f1f90e7c014f970bffabbd2fb
-updated: 2016-08-28T19:44:15.038360984+09:00
+hash: 149b93bd592f0c639ddd4777f3f496feff57278c5e5d6722e8bcd5207cb7fbfe
+updated: 2016-09-28T17:45:43.557952286+09:00
 imports:
 - name: github.com/BurntSushi/toml
   version: bbd5bb678321a0d6e58f1099321dfa73391c1b6f
@@ -17,17 +17,19 @@ imports:
   version: f390dcf405f7b83c997eac1b06768bb9f44dec18
   subpackages:
   - assert
-  - suite
   - require
+  - suite
 - name: github.com/uber-go/atomic
   version: 0c9e689d64f004564b79d9a663634756df322902
 - name: github.com/uber-go/zap
-  version: d11d2851fcabcf03421c4dbdbc3146fc74eb1035
+  version: f1942381637c11a802c356e70ab5b0a953540bb5
 - name: golang.org/x/net
-  version: b797637b7aeeed133049c7281bfa31dcc9ca42d6
+  version: f09c4662a0bd6bd8943ac7b4931e185df9471da4
   subpackages:
   - http2
   - http2/hpack
+  - idna
+  - lex/httplex
 testImports:
 - name: github.com/davecgh/go-spew
   version: 2df174808ee097f90d259e432cc04442cf60be21