--- a/glide.yaml
+++ b/glide.yaml
@@ -8,11 +8,10 @@ import:
   - payload
   - payload/badge
   - push
-- package: github.com/Sirupsen/logrus
-  version: v0.10.0
 - package: github.com/mercari/gcm
   version: 987b1dc4ce9034b698395d35de1aadf999388b8f
 - package: github.com/fukata/golang-stats-api-handler
   version: e7ee1630fdb679c86ec8e3d0755e3576675fdb10
 - package: github.com/stretchr/testify
   version: v1.1.3
+- package: github.com/uber-go/zap