--- a/glide.yaml
+++ b/glide.yaml
@@ -15,3 +15,4 @@ import:
 - package: github.com/stretchr/testify
   version: v1.1.3
 - package: github.com/uber-go/zap
+  version: f1942381637c11a802c356e70ab5b0a953540bb5