use github.com/uber-go/zap instead of github.com/Sirupsen/logrus for access/error logs. This breaks compatibility about log format and achieves fully json logging.

zap is very complex to use (ex: omitempty), but slightly faster than other loggers. This is a benchmark before/after this changes for `gaurun.LogPush`.

**before**

```
go test github.com/mercari/gaurun/gaurun -benchmem -benchtime 5s -bench .   
BenchmarkLogPushIOSOmitempty-4   	  500000	     12257 ns/op	    1816 B/op	      18 allocs/op
BenchmarkLogPushIOSFull-4        	  500000	     13893 ns/op	    2008 B/op	      18 allocs/op
```

**after**

```
go test github.com/mercari/gaurun/gaurun -benchmem -benchtime 5s -bench .     
BenchmarkLogPushIOSOmitempty-4   	 1000000	      6060 ns/op	     945 B/op	       4 allocs/op
BenchmarkLogPushIOSFull-4        	 1000000	      6303 ns/op	     945 B/op	       4 allocs/op
```

2 allocs for zap logging. 2 allocs for `ptime`'s format conversion.

**EDIT:**

After removing format conversion and computing the value instead.

```
go test github.com/mercari/gaurun/gaurun -benchmem -benchtime 5s -bench . 
BenchmarkLogPushIOSOmitempty-4   	 2000000	      4574 ns/op	     929 B/op	       2 allocs/op
BenchmarkLogPushIOSFull-4        	 1000000	      5030 ns/op	     929 B/op	       2 allocs/op
```

*change list*

- log format change
 - `[info] {"message": "...", ...}` to `{"level":"info","message":"..."...}`
- time format
 - `2015/08/04 15:32:35 JST` (local time) to `2016-08-30T01:18:04Z` (UTC)
-