@kazegusuri 2016-08-30
current time format is `t.Format("2006/01/02 15:04:05 MST")` such as `2015/08/04 15:32:35 JST` at local time. This PR changes the format to `t.Format(time.RFC3339)` such as `2016-08-30T01:18:04Z` at UTC.  If we need to keep time format, I will fix with zap.TimeFormatter.