@cubicdaiya 2016-09-23
@kazegusuri How is the progress?